Reformatting due to my changing of Editor.

+++ rnnparser/RecursiveNN/recursiveNN/math.py
@@ -3,7 +3,7 @@ import numpy as np
 
 def NormalizedMatrix(mat, l1norm):
     return l1norm*mat/np.abs(mat).sum()
-    
+
 def SimplifyIfScalar(arr):
     if np.product(arr.shape) ==1 :
         return arr.dtype.type(arr)
@@ -16,41 +16,44 @@ def ArrayOrScala(arr):
 
 def IsScalar(x):
     try:
-        m=int(x.val)
+        m=float(x.val)
         return True
     except:
         pass
     return False
+
 def IsVector(x):
     try:
         if len(x.val.shape)==1:
             return True
         m,n=x.val.shape
+        #m,n=x.val.shape[0],x.val.shape[1]
         return (m==1 and n>1) or (m>1 and n==1)
     except:
         pass
-    return False    
+    return False
 def IsMatrix(x):
     try:
+        #m,n=x.val.shape[0],x.val.shape[1]
         m,n=x.val.shape
         return m>1 and n>1
     except:
         pass
     return False
-    
-def IsZero(var):    
-    try : 
+
+def IsZero(var):
+    try :
         return np.all(var.val==0)
     except:
         pass
     return False
-def IsAllOne(var):         
-    try : 
+def IsAllOne(var):
+    try :
         return np.all(var.val==1.0)
     except:
         pass
     return False
-def IsIdentity(var):     
+def IsIdentity(var):
     try :
         if not isinstance(var.val, np.ndarray) and var.val==1.0:
             return True
@@ -63,4 +66,3 @@ def IsIdentity(var):
     except:
         pass
     return False
-
